#209: L'identifiant de session d'importation introduit par [9903] est étendu dans la fonction '''inc_admin''' à un identifiant de session d'administration, afin d'empêcher plus généralement deux connexions dans l'espace privé lorsqu'une opération lourde a lieu (actuellement: mise à jour de SPIP,  importation/fusion d'archives, conversion de caractères par SQL ou par PHP, destruction des statistiques ou de la base). Grace à ça, les deux verrous MySQL pour les conversions de caractères passent à la trappe. Plus que 3 verrous MySQL à éliminer.
git-svn-id: svn://trac.rezo.net/spip/spip@10069 caf5f3e8-d4fe-0310-bb3e-c32d5e47d55d

+++ ecrire/install/etape_3.php
@@ -97,7 +97,9 @@ function install_bases(){
 		$finsert('spip_meta', '(nom, valeur,impt)', "('version_installee', '$spip_version','non')");
 	} else {
 
-	  creer_base($server_db); // pour recrer les tables disparues au besoin
+	  // pour recreer les tables disparues au besoin
+	  creer_base($server_db); 
+
 	  $r = $fquery ("SELECT valeur FROM spip_meta WHERE nom='version_installee'");
 	  if ($r) $r = $ffetch($r);
 	  if ($r) $version_installee = (double) $r['valeur'];
@@ -138,8 +140,9 @@ function install_bases(){
 		
 		$result_ok = $finsert("spip_meta", "(nom, valeur)", "('nouvelle_install', '1')");
 	} else {
-	  // en cas de reinstall sur mise a jour mal passee
-		@$fquery("DELETE FROM spip_meta WHERE nom='import_all'");
+	  // eliminer la derniere operation d'admin mal terminee
+	  // notamment la mise a jour 
+		@$fquery("DELETE FROM spip_meta WHERE nom='import_all' OR  nom='admin'");
 		$result_ok = @$fquery("SELECT COUNT(*) FROM spip_meta");
 	}
